project: drop Makefile in favor of Taskfile
task (https://github.com/go-task/task) is a task runner written in Go
and released via the MIT license. for those reasons, it is being favored
over make for this project.

this change also switches to including wayland support as well as using
release flags in the default build command.